tests: move resourcewatcher tests into same package
Allows us to keep package-private methods private,
and is more idiomatic.
